Fix: Sales Order decimal quantity — parseInt → parseFloat
User feedback: 'qty 7.5 x unit price 800 hobe 6000 taka, but count
korche 5600, maney decimal count korche na'

ROOT CAUSE: All quantity fields used parseInt() which truncates
decimals:
  parseInt('7.5') → 7  (WRONG — should be 7.5)
  7 × 800 = 5600   (WRONG — should be 6000)

FIX: Changed parseInt → parseFloat for ALL sales order quantity
calculations. Also changed DB schema from Int to Float.

Changes:

1. Prisma schema (prisma/schema.prisma):
   - SalesOrderItem.quantity: Int → Float
   - SalesMakingEntry.quantity: Int → Float

2. Runtime migration (src/app/api/admin/migrate-schema/route.ts):
   SQLite doesn't support ALTER COLUMN TYPE, so for each table:
   a. Add new REAL column (quantity_new)
   b. Copy data: UPDATE ... SET quantity_new = CAST(quantity AS REAL)
   c. Drop old INTEGER column
   d. Rename quantity_new → quantity
   Added 8 migration steps (4 for SalesOrderItem + 4 for SalesMakingEntry)

3. Frontend (src/app/page.tsx) — 8 locations changed:
   - Line 1508: totalQty calculation
   - Line 1509: totalAmount calculation
   - Line 1530: API payload (items.map quantity)
   - Line 1530: API payload (makingEntries.map quantity)
   - Line 4709: subTotal calculation
   - Line 4710: makingTotal calculation
   - Line 4831: itemBaseTotal display
   - Line 4832: itemMakingTotal display
   - Line 4846: meTotal display
   - Line 1759: sales return quantity

4. API (src/app/api/sales-orders/route.ts) — 4 locations changed:
   - Line 81: stock aggregation
   - Line 109: SalesOrderItem create
   - Line 116: SalesMakingEntry create
   - Line 182: incentive calculation

Net effect: quantity 7.5 × price 800 = 6000.00 (correct!)
Before: parseInt('7.5') = 7 → 7 × 800 = 5600 (wrong)

IMPORTANT: After deploy, user must run migration:
  POST https://dfcl-inv.vercel.app/api/admin/migrate-schema?token=DFCL_RESCUE_2026
This will change the DB column type from INTEGER to REAL so decimal
values can be stored.

+++ src/app/api/sales-orders/route.ts
@@ -78,7 +78,7 @@ export async function POST(request: NextRequest) {
     const aggregated = new Map<string, number>();
     for (const item of items as any[]) {
       const key = `${item.itemId}|${entityId}`;
-      aggregated.set(key, (aggregated.get(key) || 0) + (parseInt(item.quantity) || 1));
+      aggregated.set(key, (aggregated.get(key) || 0) + (parseFloat(item.quantity) || 1));
     }
     for (const [key, totalQty] of aggregated.entries()) {
       const [itemId] = key.split('|');
@@ -106,14 +106,14 @@ export async function POST(request: NextRequest) {
           create: items.map((item: any) => ({
             itemId: item.itemId,
             entityId,
-            quantity: parseInt(item.quantity) || 1,
+            quantity: parseFloat(item.quantity) || 1,
             unitPrice: parseFloat(item.unitPrice) || 0,
             makingEntries: {
               create: (item.makingEntries || []).map((me: any) => ({
                 name: me.name || '',
                 makingInfoId: me.makingInfoId || null,
                 unitPrice: parseFloat(me.unitPrice) || 0,
-                quantity: parseInt(me.quantity) || 1,
+                quantity: parseFloat(me.quantity) || 1,
               })),
             },
           })),
@@ -179,7 +179,7 @@ export async function POST(request: NextRequest) {
         // For each item, find a matching range in any formula
         for (const item of items) {
           const unitPrice = parseFloat(item.unitPrice) || 0;
-          const quantity = parseInt(item.quantity) || 1;
+          const quantity = parseFloat(item.quantity) || 1;
           const fis = byItem.get(item.itemId) || [];
           // Sort formulas by createdAt asc, then find first matching range
           let matchedFormula: any = null;
